feat: Add operating hours creation to store creation process
