+++ owner-service/src/main/kotlin/com/carava/carwash/store/controller/StoreController.kt
@@ -3,9 +3,7 @@ package com.carava.carwash.store.controller
 import com.carava.carwash.common.dto.ApiResponse
 import com.carava.carwash.infrastructure.annotation.CurrentMemberId
 import com.carava.carwash.store.dto.*
-import com.carava.carwash.store.service.HolidayService
 import com.carava.carwash.store.service.MenuService
-import com.carava.carwash.store.service.OperatingHourService
 import com.carava.carwash.store.service.StoreService
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.*
@@ -15,8 +13,6 @@ import org.springframework.web.bind.annotation.*
 class StoreController(
     private val storeService: StoreService,
     private val menuService: MenuService,
-    private val operatingHourService: OperatingHourService,
-    private val holidayService: HolidayService,
 ) {
 
     @PostMapping
@@ -45,7 +41,7 @@ class StoreController(
         @RequestBody request: SaveOperatingHourRequestDto,
         @CurrentMemberId memberId: Long
     ) : ApiResponse<SaveOperatingHourResponseDto> {
-        val response = operatingHourService.saveOperatingHour(storeId, request, memberId)
+        val response = storeService.saveOperatingHour(storeId, request, memberId)
         return ApiResponse.success(data = response)
     }
 
@@ -55,7 +51,7 @@ class StoreController(
         @RequestBody request: CreateHolidayRequestDto,
         @CurrentMemberId memberId: Long
     ) : ApiResponse<CreateHolidayResponseDto> {
-        val response = holidayService.createHoliday(storeId, request, memberId)
+        val response = storeService.createHoliday(storeId, request, memberId)
         return ApiResponse.success(data = response)
     }
