+++ shared/src/main/kotlin/com/carava/carwash/domain/store/entity/Store.kt
@@ -44,8 +44,11 @@ class Store (
     @Column(name = "favorite_count", columnDefinition = "INT DEFAULT 0")
     var favoriteCount: Int = 0,
 
-    @OneToMany(mappedBy = "store", fetch = FetchType.LAZY, cascade = [CascadeType.ALL])
+    @OneToMany(mappedBy = "store", fetch = FetchType.LAZY, cascade = [CascadeType.ALL], orphanRemoval = true)
     var menus: MutableList<Menu> = mutableListOf(),
+
+    @OneToMany(mappedBy = "store", fetch = FetchType.LAZY, cascade = [CascadeType.ALL], orphanRemoval = true)
+    var operatingHours: MutableList<OperatingHour> = mutableListOf(),
 
 ) : BaseEntity() {
